Moved index.ts to sbComponents/index.ts
