+++ src/pages/_app.tsx
@@ -2,7 +2,7 @@ import "../styles/globals.css";
 import "../styles/app.css";
 import type { AppProps } from "next/app";
 import { storyblokInit, apiPlugin } from "@storyblok/react";
-import { storyblokComponents } from "../components";
+import { storyblokComponents } from "../components/sbComponents";
 
 storyblokInit({
   accessToken: process.env.STORYBLOK_ACCESS_TOKEN,
